build: split the flisp frontend out of libjulia into libjulia-frontend
Separate the flisp-based reference frontend (parsing, macro expansion
and lowering) from libjulia-internal into a new shared library,
libjulia-frontend, mirroring the existing libjulia-codegen split. This
is a first step towards making the frontend swappable, with the eventual
goal of bootstrapping Julia without flisp by compiling JuliaSyntax and
JuliaLowering into a standalone replacement library via juliac.

The runtime now reaches the frontend exclusively through a new
JL_FRONTEND_EXPORTED_FUNCS trampoline table in the loader (the same
mechanism used for codegen): jl_frontend_init, jl_frontend_parse,
jl_frontend_lower (renamed from jl_fl_parse/jl_fl_lower),
jl_macroexpand, the operator query functions, jl_lisp_prompt, and the
flisp profiling entry points. The library is optional at runtime: when
it is absent, fallback stubs in libjulia-internal raise an error for
parse/lower/macroexpand and conservatively return false/0 for the
operator queries, so value printing (including jl_static_show in error
paths) keeps working. In the other direction the frontend is an
ordinary client of the runtime's exported C API; macro invocation
(world-age selection, method lookup and LoadError wrapping) stays in
the runtime behind the newly exported jl_invoke_julia_macro, and
jl_exprn, jl_new_code_info_from_ir, jl_module_next_counter, jl_log and
jl_isa_ast_node are now exported for the frontend's use.

Concretely, the flisp-dependent half of src/ast.c (ast context pool,
scheme<->julia conversion, macro expansion driver and the parse/lower
entry points) moves to src/flfrontend.c, which is compiled with
-DJL_LIBRARY_EXPORTS_FRONTEND and linked together with libflisp and the
flisp boot image into libjulia-frontend. ast.c keeps the flisp-free AST
utilities (common symbols, jl_copy_ast, jl_has_meta, jl_isa_ast_node)
and the jl_parse/jl_lower hook dispatchers. The identifier/operator-
suffix character classification used by both the parser and runtime
printing moves from src/flisp/julia_extensions.c into libsupport
(src/support/idchars.c) so that it remains available to the runtime and
to the build-time flisp host executable.
