+++ cli/trampolines/trampolines_aarch64.S
@@ -18,4 +18,5 @@ JL_RUNTIME_EXPORTED_FUNCS(XX)
 JL_RUNTIME_EXPORTED_FUNCS_WIN(XX)
 #endif
 JL_CODEGEN_EXPORTED_FUNCS(XX)
+JL_FRONTEND_EXPORTED_FUNCS(XX)
 #undef XX
