+++ cli/trampolines/trampolines_arm.S
@@ -21,4 +21,5 @@ JL_RUNTIME_EXPORTED_FUNCS(XX)
 JL_RUNTIME_EXPORTED_FUNCS_WIN(XX)
 #endif
 JL_CODEGEN_EXPORTED_FUNCS(XX)
+JL_FRONTEND_EXPORTED_FUNCS(XX)
 #undef XX
