+++ cli/trampolines/trampolines_powerpc64le.S
@@ -26,4 +26,5 @@ CNAME(name)##: ; \
 
 JL_RUNTIME_EXPORTED_FUNCS(XX)
 JL_CODEGEN_EXPORTED_FUNCS(XX)
+JL_FRONTEND_EXPORTED_FUNCS(XX)
 #undef XX
