+++ cli/trampolines/trampolines_riscv64.S
@@ -17,4 +17,5 @@
 
 JL_RUNTIME_EXPORTED_FUNCS(XX)
 JL_CODEGEN_EXPORTED_FUNCS(XX)
+JL_FRONTEND_EXPORTED_FUNCS(XX)
 #undef XX
