+++ src/support/Makefile
@@ -8,7 +8,7 @@ JCXXFLAGS_COMMON += $(CXXFLAGS) $(JL_CXXFLAGS)
 JCPPFLAGS_COMMON += $(CPPFLAGS) $(JL_CPPFLAGS)
 JLDFLAGS += $(LDFLAGS) $(JL_LDFLAGS)
 
-SRCS := hashing timefuncs ptrhash strhash operators utf8 ios htable bitvector \
+SRCS := hashing timefuncs ptrhash strhash operators utf8 idchars ios htable bitvector \
 	int2str libsupportinit arraylist strtod rle eytzinger
 ifeq ($(OS),WINNT)
 SRCS += asprintf strptime
